fixup! added views for evidence list and add new evidence
