+++ src/apps/investment-projects/apps/evidence/router.js
@@ -2,7 +2,7 @@ const router = require('express').Router()
 
 const {
   renderAddEvidence,
-} = require('./controllers')
+} = require('./controllers/index')
 router.get('/evidence', renderAddEvidence)
 
 module.exports = router
